reorganized enroll secrets table file
