+++ frontend/components/EnrollSecretTable/_styles.scss
