+++ frontend/pages/hosts/ManageHostsPage/components/EnrollSecretModal/EnrollSecretModal.tsx
@@ -3,7 +3,7 @@ import { useSelector } from "react-redux";
 import Modal from "components/Modal";
 import Button from "components/buttons/Button";
 // @ts-ignore
-import EnrollSecretTable from "components/config/EnrollSecretTable";
+import EnrollSecretTable from "components/EnrollSecretTable";
 import { ITeam } from "interfaces/team";
 import { IEnrollSecret } from "interfaces/enroll_secret";
